+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Analyzer#urlResolver is a property that exposes the analyzer's url resolver,
   for cases where more direct access to url resolution is desired.
 * Fix a situation where a warning would be reported as `[Object object]`.
+* Fix issue where inline JavaScript module import statements did not honor
+  their containing document's baseUrl; inline documents now inherit baseUrl
+  from their containing documents.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.3] - 2017-12-08